+++ rpc/client.go
@@ -1,0 +1,26 @@
+package rpc
+
+import (
+	"context"
+	"encoding/json"
+)
+
+type callCloser interface {
+	CallContext(ctx context.Context, result interface{}, method string, args ...interface{}) error
+	Close()
+}
+
+func do(ctx context.Context, call callCloser, method string, data interface{}, args ...interface{}) error {
+	var raw json.RawMessage
+	err := call.CallContext(ctx, &raw, method, args...)
+	if err != nil {
+		return err
+	}
+	if len(raw) == 0 {
+		return errNotFound
+	}
+	if err := json.Unmarshal(raw, &data); err != nil {
+		return err
+	}
+	return nil
+}
