+++ rpc/contract_test.go
