+++ rpc/events_test.go
